requirements.txt
deepspeed
peft
torchkeras
bitsandbytes
accelerate
transformers==4.30
modelscope
xformers